refactor(ai): split near-gate planners and share destination plan shells
Extract conquest stalled/destination scoring parts and peer-peace topic
libraries under a thin barrel so AC-5 near-gate files stay well under
750 lines; store province/owner lists on PhaseDestinationResult for the
military/naval plan type shells without changing public type names.
